renamed camelCased files according to pylint
