+++ src/fritzbox_connection_uptime.py
@@ -21,7 +21,7 @@
 import sys
 from fritzconnection.lib.fritzstatus import FritzStatus
 from fritzconnection.core.exceptions import FritzConnectionException
-from FritzboxConfig import FritzboxConfig
+from fritzbox_config import FritzboxConfig
 
 class FritzboxConnectionUptime:
   __connection = None
