+++ src/fritzbox_link_saturation.py
@@ -20,7 +20,7 @@
 """
 
 import sys
-from FritzboxInterface import FritzboxInterface
+from fritzbox_interface import FritzboxInterface
 
 PAGE = 'data.lua'
 PARAMS = {'xhr':1, 'lang':'de', 'page':'netMoni', 'xhrId':'updateGraphs', 'useajax':1, 'no_sidrenew':None}
